fix: useSocket 라우팅 시점에 호출하도록 수정
