--- frontend/src/components/DashBoard/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './DashBoard';
