--- frontend/src/pages/ChannelManager/ChannelManager.stories.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-import { storiesOf } from '@storybook/react';
-import ChannelManager from '@/pages/ChannelManager';
-
-storiesOf('ChannelManager', module).add('default', () => <ChannelManager />);
